requirements.txt
moviepy==1.0.3
openai==0.27.4
openai-whisper==20230308
pydub==0.25.1
pytube==12.1.2
python-telegram-bot==13.7

